move the rest of logger code
