+++ packages/@ionic/cli-framework-output/package.json
@@ -26,7 +26,11 @@
   "dependencies": {
     "debug": "^4.0.0",
     "log-update": "^4.0.0",
-    "tslib": "1.11.2"
+    "slice-ansi": "^4.0.0",
+    "string-width": "^4.1.0",
+    "strip-ansi": "^6.0.0",
+    "tslib": "1.11.2",
+    "wrap-ansi": "^7.0.0"
   },
   "devDependencies": {
     "@ionic/utils-stream": "3.1.3",
@@ -34,6 +38,8 @@
     "@types/inquirer": "0.0.43",
     "@types/jest": "^25.1.0",
     "@types/node": "~10.17.13",
+    "@types/slice-ansi": "^4.0.0",
+    "@types/wrap-ansi": "^3.0.0",
     "jest": "^25.1.0",
     "jest-cli": "^26.0.1",
     "lint-staged": "^10.0.2",
