+++ packages/@ionic/cli-framework-output/src/index.ts
@@ -1,4 +1,5 @@
 export * from './colors';
+export * from './format';
 export * from './logger';
 export * from './output';
 export * from './tasks';
