+++ packages/@ionic/cli-framework/package.json
@@ -32,12 +32,8 @@
     "lodash": "^4.17.5",
     "minimist": "^1.2.0",
     "rimraf": "^3.0.0",
-    "slice-ansi": "^4.0.0",
-    "string-width": "^4.1.0",
-    "strip-ansi": "^6.0.0",
     "tslib": "1.11.2",
     "untildify": "^4.0.0",
-    "wrap-ansi": "^7.0.0",
     "write-file-atomic": "^3.0.0"
   },
   "devDependencies": {
@@ -46,8 +42,6 @@
     "@types/lodash": "^4.14.104",
     "@types/minimist": "^1.2.0",
     "@types/node": "~10.17.13",
-    "@types/slice-ansi": "^4.0.0",
-    "@types/wrap-ansi": "^3.0.0",
     "@types/write-file-atomic": "^3.0.0",
     "jest": "^25.1.0",
     "jest-cli": "^26.0.1",
